Update router.php script.
Prefix call to `theme_render_template($page)`with `$docroot`

This causes the `include <x>` call in `theme_render_template` to become an absolute import.

The `include <x>` call worked in most cases in the past. It was a relative include. PHP looks for relative includes in 3 places:
* The `include_path` dir
* The __DIR__ of the including file
* The cwd

However, in the RCT styleguide instance, none of the above were == $docroot.

Now we force render an absolute file path. This matches the behaviour of the twig renderer.